@@ -1,14 +1,15 @@
 package uk.org.sucu.tatupload2;
 
-import java.util.HashMap;
-
 import android.content.Context;
 import android.os.Bundle;
-import android.support.v4.app.Fragment;
-import android.support.v4.app.FragmentActivity;
-import android.support.v4.app.FragmentTransaction;
 import android.view.View;
 import android.widget.TabHost;
+
+import java.util.HashMap;
+
+import androidx.fragment.app.Fragment;
+import androidx.fragment.app.FragmentActivity;
+import androidx.fragment.app.FragmentTransaction;
 
 public class TabManager implements TabHost.OnTabChangeListener{
 	private final FragmentActivity mActivity;
