@@ -2,11 +2,11 @@ package uk.org.sucu.tatupload2.activity;
 
 import android.content.Intent;
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
 import android.text.method.LinkMovementMethod;
 import android.view.View;
 import android.widget.TextView;
 
+import androidx.appcompat.app.AppCompatActivity;
 import uk.org.sucu.tatupload2.R;
 import uk.org.sucu.tatupload2.Settings;
 import uk.org.sucu.tatupload2.parse.Parameters;
